Comments applied
function documentions added
