+++ catkin_ws/src/kpr_launch/launch/robot.launch
@@ -30,7 +30,7 @@
 			<param name="max_velocity" type="double" value="$(arg max_velocity)"/>
 			
 		</node>
-		<node name="valves" pkg="kpr_testing" type="valves.py" output="screen" launch-prefix="gnome-terminal -x"/>
+		<node name="io_pins" pkg="kpr_testing" type="io_pins.py" output="screen" launch-prefix="gnome-terminal -x"/>
 	</group>
 
 	<group unless="$(arg testing)">
