+++ catkin_ws/src/kpr_testing/package.xml
@@ -2,13 +2,10 @@
 <package>
 	<name>kpr_testing</name>
 	<version>0.0.0</version>
-	<description>The kpr_testing package</description>
+	<description>A package for testing the robots io-pins of the 'Tomatenplukkers' project</description>
 
-	<!-- One maintainer tag required, multiple allowed, one person per tag --> 
-	<!-- Example:  -->
-	<!-- <maintainer email="jane.doe@example.com">Jane Doe</maintainer> -->
-	<maintainer email="pascalle@todo.todo">pascalle</maintainer>
-
+	<maintainer email="pascalle1994@hotmail.com">pascalle</maintainer>
+	<author email="pascalle1994@hotmail.com">Jane Doe</author>
 
 	<!-- One license tag required, multiple allowed, one license per tag -->
 	<!-- Commonly used license strings: -->
@@ -16,29 +13,9 @@
 	<license>TODO</license>
 
 
-	<!-- Url tags are optional, but mutiple are allowed, one per tag -->
-	<!-- Optional attribute type can be: website, bugtracker, or repository -->
-	<!-- Example: -->
-	<!-- <url type="website">http://wiki.ros.org/kpr_testing</url> -->
+	<url type="repository">https://github.com/ChielBruin/tomatenplukkers</url>
 
 
-	<!-- Author tags are optional, mutiple are allowed, one per tag -->
-	<!-- Authors do not have to be maintianers, but could be -->
-	<!-- Example: -->
-	<!-- <author email="jane.doe@example.com">Jane Doe</author> -->
-
-
-	<!-- The *_depend tags are used to specify dependencies -->
-	<!-- Dependencies can be catkin packages or system dependencies -->
-	<!-- Examples: -->
-	<!-- Use build_depend for packages you need at compile time: -->
-	<!--   <build_depend>message_generation</build_depend> -->
-	<!-- Use buildtool_depend for build tool packages: -->
-	<!--   <buildtool_depend>catkin</buildtool_depend> -->
-	<!-- Use run_depend for packages you need at runtime: -->
-	<!--   <run_depend>message_runtime</run_depend> -->
-	<!-- Use test_depend for packages you need only for testing: -->
-	<!--   <test_depend>gtest</test_depend> -->
 	<buildtool_depend>catkin</buildtool_depend>
 	<build_depend>rospy</build_depend>
 	<run_depend>rospy</run_depend>
@@ -46,9 +23,6 @@
 	<run_depend>message_runtime</run_depend>
 
 
-	<!-- The export tag contains other, unspecified, tags -->
-	<export>
-	<!-- Other tools can request additional information be placed here -->
+	<export>	</export>
 
-	</export>
 </package>
